restructure package loading
